TSM-38 - home page *fixed unit test problem

+++ trunk/tsm/src/test/java/com/tech4d/tsm/dao/IntegrationTestEventDao.java
@@ -130,7 +130,7 @@ public class IntegrationTestEventDao extends OpenSessionInViewIntegrationTest{
         			earlier.getCreated().after(event.getCreated()));
         	earlier = event;
         }        
-        assertEquals(2, eventDao.getTotalCount());
+        assertEquals(2, (int)eventDao.getTotalCount());
 
     }
